pulled trends apart
