+++ libs/tools/trend_utils/__init__.py
@@ -1,0 +1,4 @@
+""" trend utilities """
+from .trend_of_periods import get_lines_from_period
+from .analysis import generate_analysis
+from .line_utils import filter_nearest_to_signal, consolidate_lines
